rename: rename RegionCountMarkersContainer to LowZoomMarkerContainer
[#811]
