+++ frontend/src/components/google-maps/marker/LowZoomMarkerContainer/components/RegionCountMarker.style.ts
