+++ frontend/src/components/google-maps/marker/LowZoomMarkerContainer/components/RegionCountMarker.tsx
