+++ frontend/src/components/google-maps/marker/LowZoomMarkerContainer/components/RegionCountMarkerRenderer.tsx
