+++ frontend/src/components/google-maps/marker/LowZoomMarkerContainer/hooks/useRenderRegionCountMarker.tsx
