+++ frontend/src/components/google-maps/marker/LowZoomMarkerContainer/index.ts
@@ -1,0 +1,3 @@
+import LowZoomMarkerContainer from './LowZoomMarkerContainer';
+
+export default LowZoomMarkerContainer;
