--- frontend/src/components/google-maps/marker/RegionCountMarkersContainer/index.ts
@@ -1,3 +1,0 @@
-import RegionCountMarkersContainer from './RegionCountMarkersContainer';
-
-export default RegionCountMarkersContainer;
